+++ front/temporal/scrub_workspace/activities.ts
