+++ front/temporal/scrub_workspace/config.ts
